refactor(sa): migrate pipeline stages from SACalculator to lf.sa namespace
Completes steps 8 + 9 of the SA refactor plan. Moves all remaining
pipeline stages (risk_weights, FCSM, life insurance, guarantee
substitution, currency mismatch, due diligence, defaulted) from
SACalculator private methods to the lf.sa Polars namespace — aligning
SA with engine/irb and engine/slotting.

SACalculator is now a thin orchestrator whose get_sa_result_bundle /
calculate_unified / calculate_branch methods read as fluent chains of
lf.sa.* calls. Direct tests against the old _apply_* private methods
are rewritten to use the namespace API.
