+++ tests/unit/crm/test_simple_method.py
@@ -32,6 +32,7 @@ from decimal import Decimal
 import polars as pl
 import pytest
 
+import rwa_calc.engine.sa.namespace  # noqa: F401 - Register namespace
 from rwa_calc.contracts.config import CalculationConfig
 from rwa_calc.data.tables.crr_simple_method import (
     FCSM_RW_FLOOR,
@@ -46,7 +47,6 @@ from rwa_calc.engine.crm.simple_method import (
     compute_fcsm_columns,
     undo_sa_ead_reduction,
 )
-from rwa_calc.engine.sa.calculator import SACalculator
 
 # =============================================================================
 # Fixtures
@@ -465,8 +465,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, crr_simple_config).collect()
+        result = exposures.sa.apply_fcsm_rw_substitution(crr_simple_config).collect()
         assert result["risk_weight"][0] == pytest.approx(0.50, rel=0.001)
         assert result["pre_fcsm_risk_weight"][0] == pytest.approx(1.0)
 
@@ -486,8 +485,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, crr_simple_config).collect()
+        result = exposures.sa.apply_fcsm_rw_substitution(crr_simple_config).collect()
         assert result["risk_weight"][0] == pytest.approx(0.0, abs=1e-10)
 
     def test_fully_secured_by_cross_currency_cash_floored_20pct(self, crr_simple_config):
@@ -507,8 +505,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, crr_simple_config).collect()
+        result = exposures.sa.apply_fcsm_rw_substitution(crr_simple_config).collect()
         assert result["risk_weight"][0] == pytest.approx(0.20, rel=0.001)
 
     def test_sovereign_bond_cqs2_secured(self, crr_simple_config):
@@ -526,8 +523,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, crr_simple_config).collect()
+        result = exposures.sa.apply_fcsm_rw_substitution(crr_simple_config).collect()
         assert result["risk_weight"][0] == pytest.approx(0.20, rel=0.001)
 
     def test_corporate_bond_cqs3_secured(self, crr_simple_config):
@@ -546,8 +542,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, crr_simple_config).collect()
+        result = exposures.sa.apply_fcsm_rw_substitution(crr_simple_config).collect()
         assert result["risk_weight"][0] == pytest.approx(1.0, rel=0.001)
 
     def test_no_fcsm_columns_no_change(self, comprehensive_config):
@@ -561,8 +556,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, comprehensive_config).collect()
+        result = exposures.sa.apply_fcsm_rw_substitution(comprehensive_config).collect()
         assert result["risk_weight"][0] == pytest.approx(1.0)
 
     def test_zero_fcsm_value_no_change(self, crr_simple_config):
@@ -578,8 +572,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, crr_simple_config).collect()
+        result = exposures.sa.apply_fcsm_rw_substitution(crr_simple_config).collect()
         assert result["risk_weight"][0] == pytest.approx(1.0)
 
     def test_rwa_correctness(self, crr_simple_config):
@@ -597,8 +590,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, crr_simple_config)
+        result = exposures.sa.apply_fcsm_rw_substitution(crr_simple_config)
         result = result.with_columns(
             (pl.col("ead_final") * pl.col("risk_weight")).alias("rwa")
         ).collect()
@@ -617,8 +609,7 @@ class TestFCSMRWSubstitution:
                 "exposure_class": ["corporate"],
             }
         )
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(exposures, crr_simple_config).collect()
+        result = exposures.sa.apply_fcsm_rw_substitution(crr_simple_config).collect()
         assert result["ead_calculation_method"][0] == "simple"
 
 
@@ -740,6 +731,5 @@ class TestFCSMMixedBatch:
             currency="GBP",
         )
         with_fcsm = compute_fcsm_columns(exposures, collateral, crr_simple_config)
-        calc = SACalculator()
-        result = calc._apply_fcsm_rw_substitution(with_fcsm, crr_simple_config).collect()
+        result = with_fcsm.sa.apply_fcsm_rw_substitution(crr_simple_config).collect()
         assert result["risk_weight"][0] == pytest.approx(0.0, abs=1e-10)
